backend/requirements.txt
playwright==1.45.0
pyyaml==6.0.1
beautifulsoup4==4.12.3
llama-index-core==0.10.51
llama-index-embeddings-huggingface==0.2.0
llama-index-postprocessor-cohere-rerank
torch
transformers
fastapi==0.111.0
uvicorn
pydantic
cohere
openai==1.35.13
pandas==2.2.2
faiss-cpu
sentence-transformers==2.7.0
requests==2.32.3
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-mock==3.14.0
python-multipart==0.0.9 